% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smith_waterman.R
\name{tokenize_letters}
\alias{tokenize_letters}
\title{Tokenise text into a sequence of characters}
\usage{
tokenize_letters(x)
}
\arguments{
\item{x}{a character string of length 1}
}
\value{
a character vector with the sequence of characters in \code{x}
}
\description{
Tokenise text into a sequence of characters
}
\examples{
tokenize_letters("This function just chunks up text in letters\\nOK?")
tokenize_letters("Joske  Vermeulen")
}
\seealso{
\code{\link{strsplit}}
}
