% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smith_waterman.R
\name{tokenize_spaces_punct}
\alias{tokenize_spaces_punct}
\title{Tokenise text into a sequence of words}
\usage{
tokenize_spaces_punct(x)
}
\arguments{
\item{x}{a character string of length 1}
}
\value{
a character vector with the sequence of words in \code{x}
}
\description{
Tokenise text into a sequence of words. The function uses \code{\link{strsplit}} to split text into words
by using the [:space:] and [:punct:] character classes.
}
\examples{
tokenize_spaces_punct("This just splits. Text.alongside\\nspaces right?")
tokenize_spaces_punct("Also .. multiple punctuations or ??marks")
tokenize_spaces_punct("Joske  Vermeulen")
}
\seealso{
\code{\link{strsplit}}
}
